Add post

+++ models/user.go
@@ -5,7 +5,7 @@ import (
 )
 
 type User struct {
-	ID        uint
+	ID        uint64
 	Username  string
 	Password  string
 	Email     string
